+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandMakeProjectSpecificCatalogueNormalAgain.cs
@@ -4,10 +4,12 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using System.Collections.Generic;
 using System.Linq;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.DataExport.Data;
 using Rdmp.Core.Icons.IconProvision;
+using Rdmp.Core.Repositories;
 using Rdmp.Core.ReusableLibraryCode.Icons.IconProvision;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
@@ -17,12 +19,15 @@ namespace Rdmp.Core.CommandExecution.AtomicCommands;
 public class ExecuteCommandMakeProjectSpecificCatalogueNormalAgain : BasicCommandExecution, IAtomicCommand
 {
     private Catalogue _catalogue;
+    private List<ExtractableDataSet> _extractableDataSets;
     private ExtractableDataSet _extractableDataSet;
 
-    public ExecuteCommandMakeProjectSpecificCatalogueNormalAgain(IBasicActivateItems activator, Catalogue catalogue) :
+    public ExecuteCommandMakeProjectSpecificCatalogueNormalAgain(IBasicActivateItems activator, Catalogue catalogue, ExtractableDataSet eds) :
         base(activator)
     {
         _catalogue = catalogue;
+        _extractableDataSet = eds;
+
 
         var dataExportRepository = BasicActivator.RepositoryLocator.DataExportRepository;
         if (dataExportRepository == null)
@@ -31,20 +36,14 @@ public class ExecuteCommandMakeProjectSpecificCatalogueNormalAgain : BasicComman
             return;
         }
 
-        _extractableDataSet = dataExportRepository.GetAllObjectsWithParent<ExtractableDataSet>(catalogue)
-            .SingleOrDefault();
-
-        if (_extractableDataSet == null)
+        _extractableDataSets = dataExportRepository.GetAllObjectsWithParent<ExtractableDataSet>(catalogue).Where(eds => ProjectSpecificCatalogueManager.CanMakeCatalogueNonProjectSpecific(dataExportRepository, catalogue, eds)).ToList();
+        if (!_extractableDataSets.Any())
         {
-            SetImpossible("Catalogue is not extractable");
+            SetImpossible("Cannot make Catalogue Non-Project specific");
             return;
         }
+        
 
-        if (_extractableDataSet.Project_ID == null)
-        {
-            SetImpossible("Catalogue is not a project specific Catalogue");
-            return;
-        }
     }
 
     public override string GetCommandHelp() =>
@@ -53,15 +52,25 @@ public class ExecuteCommandMakeProjectSpecificCatalogueNormalAgain : BasicComman
     public override void Execute()
     {
         base.Execute();
-
-        _extractableDataSet.Project_ID = null;
-        _extractableDataSet.SaveToDatabase();
-
-        foreach (var ei in _catalogue.GetAllExtractionInformation(ExtractionCategory.ProjectSpecific))
+        if (_extractableDataSet is null)
         {
-            ei.ExtractionCategory = ExtractionCategory.Core;
-            ei.SaveToDatabase();
+            var projectIds = _extractableDataSets.Where(eds => eds.Project_ID != null).Select(eds => (int)eds.Project_ID);
+            var selectedProj = SelectOne<Project>(BasicActivator.RepositoryLocator.DataExportRepository.GetAllObjectsInIDList<Project>(projectIds).ToList());
+            if (selectedProj is null) return;
+            _extractableDataSet = _extractableDataSets.FirstOrDefault(eds => eds.Project_ID == selectedProj.ID);
         }
+        if (_extractableDataSet == null)
+        {
+            return;
+        }
+
+        if (_extractableDataSet.Project_ID == null)
+        {
+            SetImpossible("Catalogue is not a project specific Catalogue");
+            return;
+        }
+
+        ProjectSpecificCatalogueManager.MakeCatalogueNonProjectSpecific(BasicActivator.RepositoryLocator.DataExportRepository, _catalogue, _extractableDataSet);
 
         Publish(_catalogue);
     }